third_party/blink/web_tests/platform/generic/external/wpt/custom-elements/custom-element-registry/per-global-expected.txt

This is a testharness.js-based test.
PASS Discarding the browsing context must not change window.customElements
FAIL Navigating from the initial about:blank must not replace window.customElements assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
PASS document.open() must replace window.customElements
Harness: the test ran to completion.
